+++ src/main/java/kr/hhplus/be/server/application/payment/PaymentFacade.java
@@ -38,6 +38,6 @@ public class PaymentFacade implements PayUseCase{
 		);
 
 		// tx2: 결과 반영 + 확정/해제 + outbox
-		return command.completePayment(pending.getPaymentId(), pgResp, req.getGatewayType());
+		return command.completePayment(pending.getPaymentId(), pgResp);
 	}
 }
