+++ src/main/java/kr/hhplus/be/server/application/payment/pg/PaymentGatewayAdapter.java
@@ -1,5 +1,7 @@
 package kr.hhplus.be.server.application.payment.pg;
 
+import kr.hhplus.be.server.domain.payment.PaymentGatewayPort;
+
 public interface PaymentGatewayAdapter extends PaymentGatewayPort {
 	boolean supports(PaymentGatewayType gatewayType);
 }
