+++ src/main/java/kr/hhplus/be/server/infrastructure/persistence/payment/FakePaymentGateWayAdapter.java
@@ -2,14 +2,19 @@ package kr.hhplus.be.server.infrastructure.persistence.payment;
 
 import java.util.UUID;
 
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 import kr.hhplus.be.server.api.payment.request.PaymentGatewayRequest;
 import kr.hhplus.be.server.api.payment.response.PaymentGatewayResponse;
+
+import kr.hhplus.be.server.application.payment.pg.PaymentGatewayCancelRequest;
+import kr.hhplus.be.server.application.payment.pg.PaymentGatewayCancelResponse;
 import kr.hhplus.be.server.domain.payment.PaymentGatewayPort;
 import kr.hhplus.be.server.domain.payment.PaymentGatewayStatus;
 
 @Component
+@Primary
 public class FakePaymentGateWayAdapter implements PaymentGatewayPort {
 
 	@Override
@@ -19,5 +24,10 @@ public class FakePaymentGateWayAdapter implements PaymentGatewayPort {
 		return PaymentGatewayResponse.of("tx-" + txNum,
 			PaymentGatewayStatus.SUCCESS,
 			request.getAmount(), null, null);
+	}
+
+	@Override
+	public PaymentGatewayCancelResponse cancel(PaymentGatewayCancelRequest request) {
+		return null;
 	}
 }
